+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -597,6 +597,8 @@ public class SentryOptions {
 
   private @NotNull ISocketTagger socketTagger = NoOpSocketTagger.getInstance();
 
+  private @Nullable String profilingTracesDirPath;
+
   /**
    * Configuration options for Sentry Build Distribution. NOTE: Ideally this would be in
    * SentryAndroidOptions, but there's a circular dependency issue between sentry-android-core and
@@ -2104,11 +2106,23 @@ public class SentryOptions {
    * @return the profiling traces dir. path or null if not set
    */
   public @Nullable String getProfilingTracesDirPath() {
+    if (profilingTracesDirPath != null && !profilingTracesDirPath.isEmpty()) {
+      return dsnHash != null
+          ? new File(profilingTracesDirPath, dsnHash).getAbsolutePath()
+          : profilingTracesDirPath;
+    }
+
     final String cacheDirPath = getCacheDirPath();
+
     if (cacheDirPath == null) {
       return null;
     }
+
     return new File(cacheDirPath, "profiling_traces").getAbsolutePath();
+  }
+
+  public void setProfilingTracesDirPath(final @Nullable String profilingTracesDirPath) {
+    this.profilingTracesDirPath = profilingTracesDirPath;
   }
 
   /**
@@ -3343,6 +3357,18 @@ public class SentryOptions {
 
     if (options.isEnableLogs() != null) {
       getLogs().setEnabled(options.isEnableLogs());
+    }
+
+    if (options.getProfileSessionSampleRate() != null) {
+      setProfileSessionSampleRate(options.getProfileSessionSampleRate());
+    }
+
+    if (options.getProfilingTracesDirPath() != null) {
+      setProfilingTracesDirPath(options.getProfilingTracesDirPath());
+    }
+
+    if (options.getProfileLifecycle() != null) {
+      setProfileLifecycle(options.getProfileLifecycle());
     }
   }